cooperative_tsn/src/gRPC/pyproject.toml
[tool.poetry]
name = "grpc"
version = "0.1.0"
description = ""
authors = ["Jannusch Bigge <jannusch.bigge@mailbox.tu-dresden.de>"]
license = "GPL-2.0 or later"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
grpcio = "1.65.4"
grpcio-tools = "1.65.4"
protobuf = "5.27.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
